declare much of the old code deprecated and break out into seperate modules
