+++ graviton/dryrun.py
@@ -1,10 +1,9 @@
 import base64
 import binascii
 from contextlib import redirect_stdout
-from dataclasses import dataclass
 import io
 import string
-from typing import Any, Dict, List, Optional, Union
+from typing import Any, Dict, List, Union
 
 from algosdk.future import transaction
 from algosdk.encoding import encode_address, msgpack_encode
@@ -15,35 +14,16 @@ from algosdk.v2client.models import (
     ApplicationParams,
     ApplicationStateSchema,
     Account,
-    TealKeyValue,
 )
 
+from .deprecated_dry_run import DeprecatedDryRun
+from . import models
 
-ZERO_ADDRESS = encode_address(bytes(32))
 PRINTABLE = frozenset(string.printable)
 
-
-@dataclass
-class LSig:
-    """Logic Sig program parameters"""
-
-    args: Optional[List[bytes]] = None
-
-
-@dataclass
-class App:
-    """Application program parameters"""
-
-    creator: str = ZERO_ADDRESS
-    round: Optional[int] = None
-    app_idx: int = 0
-    on_complete: int = 0
-    args: Optional[List[Union[bytes, str]]] = None
-    accounts: Optional[List[Union[str, Account]]] = None
-    global_state: Optional[List[TealKeyValue]] = None
-
-
 # ### LIGHTWEIGHT ASSERTIONS FOR RE-USE ### #
+
+
 def _msg_if(msg):
     return "" if msg is None else f": {msg}"
 
@@ -222,303 +202,8 @@ def assert_local_state_contains(addr, delta_value, txn_index, txns_res, msg=None
         _fail(msg)
 
 
-class DryrunTestCaseMixin:
-    """
-    Mixin class for unittest.TestCase
-
-    Expects self.algo_client to be initialized in TestCase.setUp
-    """
-
-    def assertPass(
-        self,
-        prog_drr_txns,
-        lsig=None,
-        app=None,
-        sender=ZERO_ADDRESS,
-        txn_index=None,
-        msg=None,
-    ):
-        """
-        Asserts that all programs pass.
-        By default it uses logic sig mode with args passed in lsig object.
-        If app is set then application call is made
-
-        Args:
-            prog_drr_txns (bytes, str, dict, list): program to run, dryrun response object or list of transactions
-            lsig (dict, LSig): logic sig program additional parameters
-            app (dict, App): app program additional parameters
-            sender (str): txn sender
-            txn_index (int): txn result index to assert in
-
-        Raises:
-            unittest.TestCase.failureException: if not passed
-            TypeError: program is not bytes or str
-        """
-        txns_res = self._checked_request(prog_drr_txns, lsig, app, sender)
-        assert_pass(txn_index, msg, txns_res)
-
-    def assertReject(
-        self,
-        prog_drr_txns,
-        lsig=None,
-        app=None,
-        sender=ZERO_ADDRESS,
-        txn_index=None,
-        msg=None,
-    ):
-        """
-        Asserts any program is rejected.
-        By default it uses logic sig mode with args passed in lsig object.
-        If app is set then application call is made
-
-        Args:
-            prog_drr_txns (bytes, str, dict, list): program to run, dryrun response object or list of transactions
-            lsig (dict, LSig): logic sig program additional parameters
-            app (dict, App): app program additional parameters
-            sender (str): txn sender
-            txn_index (int): txn result index to assert in
-
-        Raises:
-            unittest.TestCase.failureException: if not passed
-            TypeError: program is not bytes or str
-        """
-        txns_res = self._checked_request(prog_drr_txns, lsig, app, sender)
-        assert_reject(txn_index, msg, txns_res)
-
-    def assertStatus(
-        self,
-        prog_drr_txns,
-        status,
-        lsig=None,
-        app=None,
-        sender=ZERO_ADDRESS,
-        txn_index=None,
-        msg=None,
-    ):
-        """
-        Asserts that program completes with the status.
-        By default it uses logic sig mode with args passed in lsig object.
-        If app is set then application call is made
-
-        Args:
-            prog_drr_txns (bytes, str, dict, list): program to run, dryrun response object or list of transactions
-            status (str): status to assert
-            lsig (dict, LSig): logic sig program additional parameters
-            app (dict, App): app program additional parameters
-            sender (str): txn sender
-            txn_index (int): txn result index to assert in
-
-        Raises:
-            unittest.TestCase.failureException (AssetionException): if not passed
-            TypeError: program is not bytes or str
-        """
-        txns_res = self._checked_request(prog_drr_txns, lsig, app, sender)
-        assert_status(status, txn_index, msg, txns_res)
-
-    def assertNoError(
-        self,
-        prog_drr_txns,
-        lsig=None,
-        app=None,
-        sender=ZERO_ADDRESS,
-        txn_index=None,
-        msg=None,
-    ):
-        """
-        Asserts that there are no errors.
-        for example, compilation errors or application state initialization errors.
-        By default it uses logic sig mode with args passed in lsig object.
-        If app is set then application call is made
-
-        Args:
-            prog_drr_txns (bytes, str, dict, list): program to run, dryrun response object or list of transactions
-            lsig (dict, LSig): logic sig program additional parameters
-            app (dict, App): app program additional parameters
-            sender (str): txn sender
-            txn_index (int): txn result index to assert in
-
-        Raises:
-            unittest.TestCase.failureException (AssetionException): if not passed
-            TypeError: program is not bytes or str
-        """
-        drr = self._dryrun_request(prog_drr_txns, lsig, app, sender)
-        assert_no_error(drr, txn_index=txn_index, msg=msg)
-
-    def assertError(
-        self,
-        prog_drr_txns,
-        contains=None,
-        lsig=None,
-        app=None,
-        sender=ZERO_ADDRESS,
-        txn_index=None,
-        msg=None,
-    ):
-        """
-        Asserts that there are no errors.
-        for example, compilation errors or application state initialization errors.
-        By default it uses logic sig mode with args passed in lsig object.
-        If app is set then application call is made
-
-        Args:
-            prog_drr_txns (bytes, str, dict, list): program to run, dryrun response object or list of transactions
-            lsig (dict, LSig): logic sig program additional parameters
-            app (dict, App): app program additional parameters
-            sender (str): txn sender
-            txn_index (int): txn result index to assert in
-
-        Raises:
-            unittest.TestCase.failureException (AssetionException): if not passed
-            TypeError: program is not bytes or str
-        """
-
-        drr = self._dryrun_request(prog_drr_txns, lsig, app, sender)
-        assert_error(drr, contains=contains, txn_index=txn_index, msg=msg)
-
-    def assertGlobalStateContains(
-        self,
-        prog_drr_txns,
-        delta_value,
-        app=None,
-        sender=ZERO_ADDRESS,
-        txn_index=None,
-        msg=None,
-    ):
-        """
-        Asserts that execution of the program has this global delta value
-
-        Args:
-            prog_drr_txns (bytes, str, dict, list): program to run, dryrun response object or list of transactions
-            delta_value (dict): value to assert
-
-        Raises:
-            unittest.TestCase.failureException: if not passed
-            TypeError: program is not bytes or str
-        """
-
-        txns_res = self._checked_request(
-            prog_drr_txns, lsig=None, app=app, sender=sender
-        )
-        assert_global_state_contains(delta_value, txn_index, txns_res, msg=msg)
-
-    def assertLocalStateContains(
-        self,
-        prog_drr_txns,
-        addr,
-        delta_value,
-        app=None,
-        sender=ZERO_ADDRESS,
-        txn_index=None,
-        msg=None,
-    ):
-        """
-        Asserts that execution of the program has this global delta value
-
-        Args:
-            prog_drr_txns (bytes, str, dict, list): program to run, dryrun response object or list of transactions
-            addr (str): account
-            delta_value (dict): value to assert
-
-        Raises:
-            unittest.TestCase.failureException: if not passed
-            TypeError: program is not bytes or str
-        """
-
-        txns_res = self._checked_request(
-            prog_drr_txns, lsig=None, app=app, sender=sender
-        )
-        assert_local_state_contains(addr, delta_value, txn_index, txns_res, msg=msg)
-
-    def dryrun_request(self, program, lsig=None, app=None, sender=ZERO_ADDRESS):
-        """
-        Helper function for creation DryrunRequest and making the REST request
-        from program source or compiled bytes
-
-        Args:
-            program (bytes, str): program to use as a source
-            lsig (dict, LSig): logic sig program additional parameters
-            app (dict, App): app program additional parameters
-            sender (str): txn sender
-
-        Returns:
-            dict: dryrun response object
-
-        Raises:
-            TypeError: program is not bytes or str
-        """
-        drr = DryRunHelper._deprecated_dryrun(program, lsig, app, sender)
-        return self.algo_client.dryrun(drr)
-
-    def dryrun_request_from_txn(self, txns, app):
-        """
-        Helper function for creation DryrunRequest and making the REST request
-
-        Args:
-            txns (list): list of transaction to run as a group
-            app (dict, App): app program additional parameters. Only app.round and app.accounts are used.
-
-        Returns:
-            dict: dryrun response object
-
-        Raises:
-            TypeError: program is not bytes or str
-        """
-
-        if app is not None:
-            if not isinstance(app, App) and not isinstance(app, dict):
-                raise ValueError("app must be a dict or App")
-            if isinstance(app, dict):
-                app = App(**app)
-
-        rnd = None
-        accounts = None
-        apps = []
-        if app is not None:
-            if app.round is not None:
-                rnd = app.round
-            if app.accounts is not None:
-                accounts = app.accounts
-                for acc in accounts:
-                    if acc.created_apps:
-                        apps.extend(acc.created_apps)
-
-        drr = DryrunRequest(
-            txns=txns,
-            accounts=accounts,
-            round=rnd,
-            apps=apps,
-        )
-        return self.algo_client.dryrun(drr)
-
-    @staticmethod
-    def default_address():
-        """Helper function returning default zero addr"""
-        return ZERO_ADDRESS
-
-    def _dryrun_request(self, prog_drr_txns, lsig, app, sender):
-        """
-        Helper function to make a dryrun request
-        """
-        if isinstance(prog_drr_txns, dict):
-            drr = prog_drr_txns
-        elif isinstance(prog_drr_txns, list):
-            drr = self.dryrun_request_from_txn(prog_drr_txns, app)
-        else:
-            drr = self.dryrun_request(prog_drr_txns, lsig, app, sender)
-        return drr
-
-    def _checked_request(self, prog_drr_txns, lsig=None, app=None, sender=ZERO_ADDRESS):
-        """
-        Helper function to make a dryrun request and perform basic validation
-        """
-        drr = self._dryrun_request(prog_drr_txns, lsig, app, sender)
-        if drr["error"]:
-            _fail(f"error in dryrun response: {drr['error']}")
-
-        if not drr["txns"]:
-            _fail("empty response from dryrun")
-
-        return drr["txns"]
+def make_deprecated_dryrun(algo_client):
+    return DeprecatedDryRun(algo_client, DryRunHelper.dryrun_request)
 
 
 class DryRunHelper:
@@ -528,20 +213,20 @@ class DryRunHelper:
     def singleton_logicsig_request(
         cls, program: str, args: List[bytes], txn_params: Dict[str, Any]
     ):
-        return cls.dryrun_request(program, LSig(args=args), txn_params)
+        return cls.dryrun_request(program, models.LSig(args=args), txn_params)
 
     @classmethod
     def singleton_app_request(
         cls, program: str, args: List[Union[bytes, str]], txn_params: Dict[str, Any]
     ):
-        return cls.dryrun_request(program, App(args=args), txn_params)
+        return cls.dryrun_request(program, models.App(args=args), txn_params)
 
     @classmethod
     def dryrun_request(cls, program, lsig_or_app, txn_params):
         assert isinstance(
-            lsig_or_app, (LSig, App)
+            lsig_or_app, (models.LSig, models.App)
         ), f"Cannot handle {lsig_or_app} of type {type(lsig_or_app)}"
-        is_app = isinstance(lsig_or_app, App)
+        is_app = isinstance(lsig_or_app, models.App)
         if is_app:
             return cls._app_request(program, lsig_or_app, txn_params)
 
@@ -550,7 +235,7 @@ class DryRunHelper:
     @classmethod
     def _app_request(cls, program, app, txn_params):
         """TODO: out to be able to stop delegating at tis point"""
-        run_mode = cls._get_run_mode(app)
+        run_mode = models.get_run_mode(app)
         enriched = cls._prepare_app(app)
         txn = transaction.ApplicationCallTxn(**txn_params)
         return cls._prepare_app_source_request(program, enriched, run_mode, txn)
@@ -563,105 +248,10 @@ class DryRunHelper:
         return cls._prepare_lsig_source_request(program, enriched, txn)
 
     @classmethod
-    def _deprecated_dryrun(cls, program, lsig=None, app=None, sender=ZERO_ADDRESS):
-        """
-        Helper function for creation DryrunRequest object from a program.
-        By default it uses logic sig mode
-        and if app_idx / on_complete are set then application call is made
-
-        Args:
-            program (bytes, string): program to use as a source
-            lsig (dict, LSig): logic sig program additional parameters
-            app (dict, App): app program additional parameters
-
-        Returns:
-            DryrunRequest: dryrun request object
-
-        Raises:
-            TypeError: program is not bytes or str
-            ValueError: both lsig and app parameters provided or unknown type
-        """
-
-        if lsig is not None and app is not None:
-            raise ValueError("both lsig and app not supported")
-
-        if app and not isinstance(app, (App, dict)):
-            raise ValueError("app must be a dict or App")
-
-        if lsig and not isinstance(lsig, (LSig, dict)):
-            raise ValueError("lsig must be a dict or LSig")
-
-        if isinstance(app, dict):
-            app = App(**app)
-        elif isinstance(lsig, dict):
-            lsig = LSig(**lsig)
-        else:  # both are None
-            lsig = LSig()
-
-        if not isinstance(program, (bytes, str)):
-            raise TypeError("program must be bytes or str")
-
-        run_mode = cls._get_run_mode(app)
-        is_app = run_mode != "lsig"
-        lsig_or_app = app if is_app else lsig
-        txn_params = cls.sample_txn_params(sender, is_app)
-
-        if isinstance(program, str):
-            return cls.dryrun_request(program, lsig_or_app, txn_params)
-
-        assert (
-            False
-        ), f"this is unreachable  as far as I can tell - type(program)={type(program)} - (HAVE DEPRECATED program of type bytes)!!!"
-        # del app
-        # del lsig
-
-        # # in case of bytes:
-        # sources = []
-        # apps = []
-        # accounts = []
-        # rnd = None
-
-        # txn = (
-        #     transaction.ApplicationCallTxn(**txn_params)
-        #     if is_app
-        #     else transaction.PaymentTxn(**txn_params)
-        # )
-        # if run_mode != "lsig":
-        #     txns = [cls._build_appcall_signed_txn(txn, app_or_lsig)]
-        #     application = cls.sample_app(sender, app_or_lsig, program)
-        #     apps = [application]
-        #     accounts = app_or_lsig.accounts
-        #     rnd = app_or_lsig.round
-        # else:
-        #     txns = [cls._build_logicsig_txn(program, txn, app_or_lsig)]
-
-        # return DryrunRequest(
-        #     txns=txns,
-        #     sources=sources,
-        #     apps=apps,
-        #     accounts=accounts,
-        #     round=rnd,
-        # )
-
-    @classmethod
-    def _get_run_mode(cls, app):
-        run_mode = "lsig"
-        if app is not None:
-            on_complete = (
-                app.get("on_complete") if isinstance(app, dict) else app.on_complete
-            )
-            run_mode = (
-                "clearp"
-                if on_complete == transaction.OnComplete.ClearStateOC
-                else "approv"
-            )
-        return run_mode
-
-    @classmethod
     def _prepare_app(cls, app):
         # TODO: This code is smelly. Make it less so.
         if isinstance(app, dict):
-            app = App(**app)
+            app = models.App(**app)
 
         if app.app_idx is None:
             app.app_idx = 0
@@ -679,9 +269,9 @@ class DryRunHelper:
     @classmethod
     def _prepare_lsig(cls, lsig):
         if lsig is None:
-            lsig = LSig()
+            lsig = models.LSig()
         elif isinstance(lsig, dict):
-            lsig = LSig(**lsig)
+            lsig = models.LSig(**lsig)
         return lsig
 
     @classmethod
@@ -742,14 +332,6 @@ class DryRunHelper:
         if app.accounts is not None:
             txn.accounts = [a.address for a in app.accounts]
         return transaction.SignedTransaction(txn, None)
-
-    @classmethod
-    def sample_txn_params(cls, sender: str, is_app: bool):
-        sp = transaction.SuggestedParams(int(1000), int(1), int(100), "", flat_fee=True)
-        if is_app:
-            return dict(sender=sender, sp=sp, index=0, on_complete=0)
-
-        return dict(sender=sender, sp=sp, receiver=sender, amt=0)
 
     @staticmethod
     def sample_app(sender, app, program=None):
